Remove explicit table reference from Views.Tbody
The Tbody view might at some point get attached to a different table
that does not have the id "webdlmon". Handle this case correctly by
triggering the jQuery event "update" on the current element and use the
DOM propation in jQuery to let the event update reach the parent table,
which will then tell tablesorter about the change.