Make reload from browser history page work
if you re-loaded the app from a child page (such as
  dataloggers/TA_146A/graph/bw24), the app would error out because the
  collections hadn't been initilized yet.

Also added the various apache .htaccess files for use in development.
Note that everything is still highly dependent on the install location,
and you'll have to change the base URL in multiple locations. Off the
top of my head this is:
* index.html
* app.coffee
* .htaccess.*